+++ sdk/go/kubernetes/apps/v1beta1/statefulSet.go
@@ -27,6 +27,8 @@ import (
 // If the StatefulSet has not reached a Ready state after 10 minutes, it will
 // time out and mark the resource update as Failed. You can override the default timeout value
 // by setting the 'customTimeouts' option on the resource.
+//
+// Deprecated: apps/v1beta1/StatefulSet is deprecated by apps/v1/StatefulSet and not supported by Kubernetes v1.16+ clusters.
 type StatefulSet struct {
 	pulumi.CustomResourceState