+++ sdk/go/kubernetes/extensions/v1beta1/ingress.go
@@ -25,6 +25,8 @@ import (
 // If the Ingress has not reached a Ready state after 10 minutes, it will
 // time out and mark the resource update as Failed. You can override the default timeout value
 // by setting the 'customTimeouts' option on the resource.
+//
+// Deprecated: extensions/v1beta1/Ingress is deprecated by networking/v1beta1/Ingress and not supported by Kubernetes v1.20+ clusters.
 type Ingress struct {
 	pulumi.CustomResourceState