+++ tests/integration/go/go_test.go
@@ -23,9 +23,10 @@ import (
 func TestGo_Basic(t *testing.T) {
 	integration.ProgramTest(t, &integration.ProgramTestOptions{
 		Dir: "basic",
-		Dependencies: []string{
-			"github.com/pulumi/pulumi-kubernetes/sdk/v2",
-		},
+		// TODO: uncomment and delete go.mod/go.sum once apiextensions.CustomResource is available in a tagged release.
+		//Dependencies: []string{
+		//	"github.com/pulumi/pulumi-kubernetes/sdk/v2",
+		//},
 		Quick: true,
 	})
 }
@@ -33,9 +34,10 @@ func TestGo_Basic(t *testing.T) {
 func TestGo_YAML(t *testing.T) {
 	integration.ProgramTest(t, &integration.ProgramTestOptions{
 		Dir: "yaml",
-		Dependencies: []string{
-			"github.com/pulumi/pulumi-kubernetes/sdk/v2",
-		},
+		// TODO: uncomment and delete go.mod/go.sum once apiextensions.CustomResource is available in a tagged release.
+		//Dependencies: []string{
+		//	"github.com/pulumi/pulumi-kubernetes/sdk/v2",
+		//},
 		Quick: true,
 	})
 }